+++ SeikenDensetsu3/rom.hpp
@@ -7,9 +7,9 @@ namespace ROM
 
 void open(const char* fname);
 const u8* get_rom();
-size_t get_size();
+int get_size();
 
-u8  rd  (u32 addr);
-u16 rd_w(u32 addr);
+u8  rd  (int addr);
+u16 rd_w(int addr);
 
 }
